Add /-delimited data structure
This adds pathTree, a tree-based data structure
to represent /-based hierarchies.

It has three operations:

- Set(path, v):
  Sets an explicit value at the given path,
  overwriting any existing values.
- Lookup(path):
  Retrieves the value at the given path.
  If the path does not have an explicit value,
  inherits the value from the closest ancestor.
  This operation costs no allocations.
  It can be used to match HTTP handlers.
- ListByPath(path):
  Lists explicit values at or under the given path.
  This can be used to list an index of items
  under a given path.

The pathTree structure is well-suited
to be used as a matcher for HTTP handler hierarchies,
and is optimized for fast, zero-allocation lookups.

Testing:
Besides unit tests, this includes a rapid-based state machine test.
See documentation for pgregory.net/rapid for more information
but in short:

- StateMachine.Init is called once at start
- StateMachine.Check is called after each action
- Every other method on that object is treated as an action
  and is picked from randomly

This test verifies that pathTree behaves correctly
for a bunch of randomized scenarios,
comparing the result against a less efficient storage format
for this data.